+++ build.sbt
@@ -27,7 +27,7 @@ val fuzzyc2cpgVersion = "1.1.44"
 lazy val codepropertygraph = Projects.codepropertygraph
 lazy val protoBindings = Projects.protoBindings
 lazy val semanticcpg = Projects.semanticcpg
-lazy val dataflowengine = Projects.dataflowengine
+lazy val dataflowengineoss = Projects.dataflowengineoss
 lazy val cpgserver = Projects.cpgserver
 lazy val cpgvalidator = Projects.cpgvalidator
 lazy val console = Projects.console
