+++ project/Projects.scala
@@ -4,7 +4,7 @@ object Projects {
   lazy val codepropertygraph = project.in(file("codepropertygraph"))
   lazy val protoBindings = project.in(file("proto-bindings"))
   lazy val semanticcpg = project.in(file("semanticcpg"))
-  lazy val dataflowengine = project.in(file("dataflowengine"))
+  lazy val dataflowengineoss = project.in(file("dataflowengineoss"))
   lazy val cpgserver = project.in(file("cpgserver"))
   lazy val cpgvalidator = project.in(file("cpgvalidator"))
   lazy val console = project.in(file("console"))
