+++ queries/build.sbt
@@ -1,7 +1,7 @@
 name := "queries"
 
 dependsOn(Projects.semanticcpg % "compile -> compile; test -> test",
-          Projects.dataflowengine % "compile -> compile; test -> test")
+          Projects.dataflowengineoss % "compile -> compile; test -> test")
 
 libraryDependencies ++= Seq(
   "org.scalatest" %% "scalatest" % "3.0.8" % Test,
